requirements.txt
Django==3.2.3
django-extensions==3.1.3
psycopg2==2.8.6
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==3.2.0
pytest==6.2.4
pytest-django==4.3.0
pytest-factoryboy==2.1.0
